backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator>=2.0.0
google-genai>=0.2.0
slowapi>=0.1.9
Pillow>=10.0.0
pgvector==0.2.4
redis>=4.5.0  # 可选：用于向量缓存

